Ticket 49401 - Fix compiler incompatible-pointer-types warnings
Bug Description: vs->sorted was integer pointer in older versions,
    but now it's size_t pointer, this is causing compiler warnings
    during the build

Fix Description: use size_t pointers instead of integer pointers for vs->sorted and sorted2

Review By: mreynolds

(cherry picked from commit 52ba2aba49935989152010aee0d40b01d7a78432)

+++ ldap/servers/slapd/valueset.c
@@ -744,7 +744,7 @@ valueset_array_purge(const Slapi_Attr *a, Slapi_ValueSet *vs, const CSN *csn)
     int nv = 0;
     int numValues = 0;
     Slapi_Value **va2 = NULL;
-    int *sorted2 = NULL;
+    size_t *sorted2 = NULL;
 
     /* Loop over all the values freeing the old ones. */
     for(i = 0; i < vs->num; i++)
@@ -814,7 +814,7 @@ valueset_array_purge(const Slapi_Attr *a, Slapi_ValueSet *vs, const CSN *csn)
         if(vs->sorted) {
             /* Let's allocate va2 and sorted2 */
             va2 = (Slapi_Value **) slapi_ch_malloc( (numValues + 1) * sizeof(Slapi_Value *));
-            sorted2 = (int *) slapi_ch_malloc( (numValues + 1)* sizeof(int));
+            sorted2 = (size_t *) slapi_ch_malloc( (numValues + 1)* sizeof(size_t));
         }
 
         /* I is the index for the *new* va2 array */
@@ -868,7 +868,7 @@ valueset_array_purge(const Slapi_Attr *a, Slapi_ValueSet *vs, const CSN *csn)
 
     /* We still have values but not sorted array! rebuild it */
     if(vs->num > VALUESET_ARRAY_SORT_THRESHOLD && vs->sorted == NULL) {
-        vs->sorted = (int *) slapi_ch_malloc( vs->max* sizeof(int));
+        vs->sorted = (size_t *) slapi_ch_malloc( vs->max* sizeof(size_t));
         valueset_array_to_sorted(a, vs);
     }
 #ifdef DEBUG
